requirements.txt
torch>=2.0.0
torchvision>=0.15.0
git+https://github.com/openai/CLIP.git
Pillow>=10.0.0
